# api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for TransformDefinitionType.
const (
	Javascript TransformDefinitionType = "javascript"
)

// Valid indicates whether the value is a known member of the TransformDefinitionType enum.
func (e TransformDefinitionType) Valid() bool {
	switch e {
	case Javascript:
		return true
	default:
		return false
	}
}

// EmptyResponse defines model for EmptyResponse.
type EmptyResponse = map[string]interface{}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string  `json:"error"`
	ReqID *string `json:"reqID,omitempty"`
}

// IngestRequest defines model for IngestRequest.
type IngestRequest = map[string]interface{}

// TransformDefinition defines model for TransformDefinition.
type TransformDefinition struct {
	Id         *string                 `json:"id,omitempty"`
	Script     *string                 `json:"script,omitempty"`
	ScriptPath *string                 `json:"scriptPath,omitempty"`
	Type       TransformDefinitionType `json:"type"`
}

// TransformDefinitionType defines model for TransformDefinition.Type.
type TransformDefinitionType string

// TransformRequest defines model for TransformRequest.
type TransformRequest struct {
	Body                string               `json:"body"`
	Headers             *map[string][]string `json:"headers,omitempty"`
	TransformDefinition TransformDefinition  `json:"transformDefinition"`
}

// TransformResponse defines model for TransformResponse.
type TransformResponse struct {
	Body    *string              `json:"body,omitempty"`
	Headers *map[string][]string `json:"headers,omitempty"`
}

// IngestJSONRequestBody defines body for Ingest for application/json ContentType.
type IngestJSONRequestBody = IngestRequest

// TransformJSONRequestBody defines body for Transform for application/json ContentType.
type TransformJSONRequestBody = TransformRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Ingest a webhook message
	// (POST /ingest/{sourceSlug})
	Ingest(w http.ResponseWriter, r *http.Request, sourceSlug string)
	// Get Prometheus metrics
	// (GET /metrics)
	Metrics(w http.ResponseWriter, r *http.Request)
	// Transform a message
	// (POST /transform)
	Transform(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Ingest a webhook message
// (POST /ingest/{sourceSlug})
func (_ Unimplemented) Ingest(w http.ResponseWriter, r *http.Request, sourceSlug string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get Prometheus metrics
// (GET /metrics)
func (_ Unimplemented) Metrics(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Transform a message
// (POST /transform)
func (_ Unimplemented) Transform(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// Ingest operation middleware
func (siw *ServerInterfaceWrapper) Ingest(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "sourceSlug" -------------
	var sourceSlug string

	err = runtime.BindStyledParameterWithOptions("simple", "sourceSlug", chi.URLParam(r, "sourceSlug"), &sourceSlug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "sourceSlug", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Ingest(w, r, sourceSlug)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Metrics operation middleware
func (siw *ServerInterfaceWrapper) Metrics(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Metrics(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Transform operation middleware
func (siw *ServerInterfaceWrapper) Transform(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Transform(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/ingest/{sourceSlug}", wrapper.Ingest)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/metrics", wrapper.Metrics)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/transform", wrapper.Transform)
	})

	return r
}